- patches.suse/ocfs2-Implementation-of-local-and-global-quota-file.patch:   ocfs2: Implementation of local and global quota file handling   (fate#302681). - fixed 64-bit division
suse-commit: 9983a51fd7106dcbad97229775165857d2416c3b

+++ fs/ocfs2/quota_local.c
@@ -88,7 +88,7 @@ static int ol_dqblk_chunk_off(struct super_block *sb, int c, loff_t off)
 
 	return ((off >> sb->s_blocksize_bits) -
 			ol_quota_chunk_block(sb, c) - 1) * epb
-	       + (off & ((1 << sb->s_blocksize_bits) - 1)) /
+	       + ((unsigned int)(off & ((1 << sb->s_blocksize_bits) - 1))) /
 		 sizeof(struct ocfs2_local_disk_dqblk);
 }
